move to ecs from structure
